aprend_nao_superv_agrupamento/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
joblib==1.4.2
kiwisolver==1.4.8
kmodes==0.12.2
matplotlib==3.10.1
narwhals==1.36.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.0.0
